+++ CHANGES.md
@@ -17,6 +17,7 @@
 - The **Deposit Keep Amount** and **Trainer Protection Value** sliders are now CVar sliders, complete with a companion check button.
   - _**Author's Note**: The check button now serves as the primary method to toggle functionality on and off. Setting the sliders to 0 will no longer disable their effects. For example, setting the 'Deposit Keep Amount' to 0 will cause the addon to deposit all of your gold into the Warbank and prevent any withdrawals._
 - The tooltip for the chevrons on the World Event Queue button now updates immediately upon clicking, eliminating the need to leave and re-enter the region for the tooltip to refresh.
+- Quests can now be quickly abandoned from the Objective Tracker.
 ### Deprecated
 _No changes_
 ### Removed